requirements.txt
ImageHash==4.0
numpy==1.17.4
Pillow==6.2.1
PyWavelets==1.1.1
scipy==1.4.0
six==1.13.0
